Fixed fraction cent rounding and moved cents editor
