+++ meteor-app/imports/ui/events/lister.js
@@ -5,6 +5,7 @@ import Events, { defaultObject } from '/imports/api/events/schema'
 import List from './list'
 import DateEditor from 'react-tabulator/lib/editors/DateEditor'
 import CONSTANTS from '/imports/api/constants'
+import { dollarInput } from '/imports/ui/utils/editors'
 
 const remove = id => Meteor.call('rm.Events', id)
 const update = form => {
@@ -44,9 +45,13 @@ const columns = [
   {
     field: 'price',
     title: 'Price',
-    editor: 'number',
-    mutatorEdit: value => value * 100,
-    formatter: cell => (cell.getValue() / 100).toLocaleString('en-AU', { style: 'currency', currency: 'AUD' })
+    editor: dollarInput,
+    mutatorEdit: value => Math.round(value * 100),
+    formatter: cell =>
+      (cell.getValue() / 100).toLocaleString('en-AU', {
+        style: 'currency',
+        currency: 'AUD'
+      })
   },
   {
     field: 'type',
